remarks fixed
