+++ src/main/java/client/CourierApi.java
@@ -1,4 +1,4 @@
-package Client;
+package client;
 
 import io.qameta.allure.Step;
 import io.restassured.response.Response;
